+++ bloom.h
@@ -19,10 +19,10 @@ struct bloom
   // These fields are part of the public interface of this structure.
   // Client code may read these values if desired. Client code MUST NOT
   // modify any of these.
-  int entries;
+  size_t entries;
   double error;
-  int bits;
-  int bytes;
+  size_t bits;
+  size_t bytes;
   int hashes;
 
   // Fields below are private to the implementation. These may go away or
@@ -60,7 +60,7 @@ struct bloom
  *     1 - on failure
  *
  */
-int bloom_init(struct bloom * bloom, int entries, double error);
+int bloom_init(struct bloom * bloom, size_t entries, double error);
 
 
 /** ***************************************************************************
@@ -80,7 +80,7 @@ int bloom_init(struct bloom * bloom, int entries, double error);
  *    -1 - bloom not initialized
  *
  */
-int bloom_check(struct bloom * bloom, const void * buffer, int len);
+int bloom_check(struct bloom * bloom, const void * buffer, size_t len);
 
 
 /** ***************************************************************************
@@ -101,7 +101,7 @@ int bloom_check(struct bloom * bloom, const void * buffer, int len);
  *    -1 - bloom not initialized
  *
  */
-int bloom_add(struct bloom * bloom, const void * buffer, int len);
+int bloom_add(struct bloom * bloom, const void * buffer, size_t len);
 
 
 /** ***************************************************************************
